https://osherove.com/tdd-kata-1
